Refactor estimator into dedicated crate
This moves the resource estimator code into its own crate, making it easier to do further refactoring and manage dependencies down the line.
