+++ .github/CODEOWNERS
@@ -11,6 +11,7 @@
 /npm @billti @cesarzc @minestarks
 /pip @billti @idavis @minestarks
 /playground @billti @cesarzc @minestarks
+/resource_estimator @billti @ivanbasov @swernli
 /samples @cesarzc @DmitryVasilevsky @sezna @swernli
 /vscode @billti @idavis @minestarks
 /wasm @billti @cesarzc @swernli
