+++ resource_estimator/src/estimates/data/physical_counts.rs
