+++ resource_estimator/src/estimates/modeling/physical_qubit.rs
