+++ resource_estimator/src/estimates/optimization/population.rs
