+++ resource_estimator/src/estimates/serialization.rs
