+++ wasm/Cargo.toml
@@ -26,6 +26,7 @@ qsls = { path = "../language_service" }
 qsc = { path = "../compiler/qsc" }
 qsc_codegen = { path = "../compiler/qsc_codegen" }
 qsc_project = { path = "../compiler/qsc_project", features = ["async"] }
+resource_estimator = { path = "../resource_estimator" }
 serde = { workspace = true, features = ["derive"]}
 serde-wasm-bindgen = { workspace = true}
 wasm-bindgen = { workspace = true}
